+++ CHANGELOG.md
@@ -38,7 +38,7 @@
 - PR #1993 Add iterator driven reduction for mean, var, std
 - PR #2220 Update Java bindings for reduction rename
 - PR #2224 implement isna, isnull, notna as dataframe functions
-
+- PR #2236 Implement drop_duplicates for Series
 
 ## Bug Fixes